+++ R/plot.r
@@ -91,6 +91,78 @@ gv <- function(x, name = NA_character_, ...) {
   UseMethod("gv", x)
 }
 
+#' Generate D2 input text to plot objects
+#'
+#' Produces text input for D2 to make a diagram of a given object, usually
+#' rendered with SVG.
+#'
+#' The D2 language is in an early stage of development (pre-v1.0), so it may be
+#' subject to changes that make it unable to use output from the current version
+#' of \code{d2}.
+#'
+#' Details of what is plotted are given in individual methods. There are
+#' expected commonalities, which are described below.
+#'
+#' The object is expected to be one of the following:
+#' \itemize{
+#'   \item an object whose elements have the same length. Examples would be
+#'   data frames, matrices, and other objects that can represent relations, with
+#'   names for the elements, and an optional name for the object itself.
+#'   \item a graph of sub-objects, each of which represent a relation as
+#'   described above, possibly with connections between the objects, and an
+#'   optional name for the graph as a whole.
+#' }
+#'
+#' Each relation is presented as a record-like shape, with the following elements:
+#' \itemize{
+#'   \item A optional header with the relation's name, and the number of (unique)
+#'   records.
+#'   \item A set of rows, one for each attribute in the relation. These rows
+#'   have the following contents:
+#'   \itemize{
+#'     \item the attribute names.
+#'     \item a depiction of the relation's (candidate) keys. Each
+#'     column represents a key, and a filled cell indicates that the attribute
+#'     in that row is in that key. The keys are given in lexical order, with
+#'     precedence given to keys with fewer attributes, and keys with attributes
+#'     that appear earlier in the original data frame's attribute order. Default
+#'     output from other package functions will thus have the primary key given
+#'     first. In the future, this will be changed to always give the primary key
+#'     first.
+#'     \item optionally, the attribute types: specifically, the first element
+#'     when passing the attribute's values into \code{\link{class}}.
+#'   }
+#' }
+#'
+#' Any foreign key references between relations are represented by one-way arrows,
+#' one per attribute in the foreign key.
+#'
+#' If the object has a name, then currently the name is not used, except as a
+#' single data frame's name. In the future, this will be used to give a name to
+#' the generated board, to make use of D2's composition features.
+#'
+#' @param x an object to be plotted.
+#' @param ... further arguments passed to or from other methods.
+#'
+#' @return A scalar character, containing text input for D2.
+#' @seealso \code{\link{d2.data.frame}} and \code{\link{d2.relation_schema}} for
+#'   individual methods.
+#'
+#' D2 language site: \url{https://d2lang.com}
+#'
+#' Playground for running online without installation:
+#' \url{https://play.d2lang.com/}
+#'
+#' Quarto extension: \url{https://github.com/data-intuitive/quarto-d2}
+#' @examples
+#' \dontrun{
+#' # simple data.frame example
+#' cat(d2(ChickWeight, "chick"))
+#' }
+d2 <- function(x, ...) {
+  UseMethod("d2", x)
+}
+
 #' Generate Graphviz input text to plot databases
 #'
 #' Produces text input for Graphviz to make an HTML diagram of a given database.
@@ -116,9 +188,9 @@ gv.database <- function(x, name = NA_character_, ...) {
     stop("name must be a length-one character")
   x_labelled <- to_labelled(x)
   x_elemented <- to_elemented(x)
-  setup_string <- gv_setup_string(name)
+  setup_string <- setup_string_gv(name)
   df_strings <- mapply(
-    relation_string,
+    relation_string_gv,
     attrs = attrs(x_elemented),
     attr_labels = attrs(x_labelled),
     keys = keys(x_elemented),
@@ -167,9 +239,9 @@ gv.relation <- function(x, name = NA_character_, ...) {
     stop("name must be a length-one character")
   x_labelled <- to_labelled(x)
   x_elemented <- to_elemented(x)
-  setup_string <- gv_setup_string(name)
+  setup_string <- setup_string_gv(name)
   df_strings <- mapply(
-    relation_string,
+    relation_string_gv,
     attrs = attrs(x_elemented),
     attr_labels = attrs(x_labelled),
     keys = keys(x_elemented),
@@ -219,7 +291,7 @@ gv.database_schema <- function(x, name = NA_character_, ...) {
     stop("name must be a length-one character")
   x_labelled <- to_labelled(x)
   x_elemented <- to_elemented(x)
-  setup_string <- gv_setup_string(name)
+  setup_string <- setup_string_gv(name)
   df_strings <- mapply(
     relation_schema_string,
     attrs = attrs(x_elemented),
@@ -266,7 +338,7 @@ gv.relation_schema <- function(x, name = NA_character_, ...) {
     stop("name must be a length-one character")
   x_labelled <- to_labelled(x)
   x_elemented <- to_elemented(x)
-  setup_string <- gv_setup_string(name)
+  setup_string <- setup_string_gv(name)
   df_strings <- mapply(
     relation_schema_string,
     attrs = attrs(x_elemented),
@@ -280,6 +352,44 @@ gv.relation_schema <- function(x, name = NA_character_, ...) {
   paste(
     setup_string,
     "",
+    df_strings,
+    teardown_string,
+    sep = "\n"
+  )
+}
+
+#' Generate D2 input text to plot relation schemas
+#'
+#' Produces text input for D2 to make a diagram of a given relation schema,
+#' usually rendered with SVG.
+#'
+#' Each relation in the schema is presented as a set of rows, one for each
+#' attribute in the relation. These rows do not include information about the
+#' attribute classes.
+#'
+#' @param x a relation schema, as given by \code{\link{relation_schema}} or
+#'   \code{\link{synthesise}}.
+#' @param name a character scalar, giving the name of the schema, if any.
+#' @inheritParams d2
+#'
+#' @return A scalar character, containing text input for D2.
+#' @seealso The generic \code{\link{d2}}.
+d2.relation_schema <- function(x, name = NA_character_, ...) {
+  if (any(names(x) == ""))
+    stop("relation schema names can not be zero characters in length")
+  x_labelled <- x
+  x_elemented <- x
+  df_strings <- mapply(
+    relation_schema_string_d2,
+    attrs = attrs(x_elemented),
+    attr_labels = attrs(x_labelled),
+    keys = keys(x_elemented),
+    name = names(x_elemented),
+    label = names(x_labelled)
+  ) |>
+    paste(collapse = "\n")
+  teardown_string <- ""
+  paste(
     df_strings,
     teardown_string,
     sep = "\n"
@@ -310,10 +420,10 @@ gv.data.frame <- function(x, name = NA_character_, ...) {
     stop("name must be non-empty")
   if (!is.character(name) || length(name) != 1)
     stop("name must be a length-one character")
-  setup_string <- gv_setup_string(name)
+  setup_string <- setup_string_gv(name)
   x_labelled <- x
   names(x_labelled) <- to_attr_name(names(x))
-  table_string <- relation_string(
+  table_string <- relation_string_gv(
     attrs = to_element_name(names(x)),
     attr_labels = colnames(x_labelled),
     keys = list(),
@@ -333,7 +443,51 @@ gv.data.frame <- function(x, name = NA_character_, ...) {
   )
 }
 
-gv_setup_string <- function(df_name) {
+#' Generate D2 input text to plot a data frame
+#'
+#' Produces text input for D2 to make a diagram of a given data frame, usually
+#' rendered with SVG.
+#'
+#' The rows in the plotted data frame include information about the attribute
+#' classes.
+#'
+#' @param x a data.frame.
+#' @param name a character scalar, giving the name of the record, if any. The
+#'   name must be non-empty, since it is also used to name the single table in
+#'   the plot. Defaults to `NA`: if left missing, it is set to "data".
+#' @inheritParams d2
+#'
+#' @return A scalar character, containing text input for Graphviz.
+#' @seealso The generic \code{\link{d2}}.
+d2.data.frame <- function(x, name = NA_character_, ...) {
+  if (is.na(name))
+    name <- "data"
+  if (name == "")
+    stop("name must be non-empty")
+  if (!is.character(name) || length(name) != 1)
+    stop("name must be a length-one character")
+
+  x_labelled <- x
+  names(x_labelled) <- names(x)
+  table_string <- relation_string_d2(
+    attrs = names(x),
+    attr_labels = colnames(x_labelled),
+    keys = list(),
+    name = name,
+    label = paste0("\"", name, "\""),
+    classes = vapply(x, \(a) class(a)[[1]], character(1)),
+    nrow = nrow(x),
+    row_name = "row"
+  )
+  teardown_string <- ""
+  paste(
+    table_string,
+    teardown_string,
+    sep = "\n"
+  )
+}
+
+setup_string_gv <- function(df_name) {
   paste0(
     "digraph ",
     if (!is.na(df_name))
@@ -344,7 +498,7 @@ gv_setup_string <- function(df_name) {
   )
 }
 
-relation_string <- function(
+relation_string_gv <- function(
   attrs,
   attr_labels,
   keys,
@@ -356,7 +510,7 @@ relation_string <- function(
 ) {
   row_name <- match.arg(row_name)
 
-  columns_string <- columns_string(
+  columns_string <- columns_string_gv(
     attrs,
     attr_labels,
     keys,
@@ -383,6 +537,36 @@ relation_string <- function(
     "\n",
     columns_label,
     "\n    </TABLE>>];"
+  )
+}
+
+relation_string_d2 <- function(
+  attrs,
+  attr_labels,
+  keys,
+  name,
+  label,
+  classes,
+  nrow,
+  row_name = c("record", "row")
+) {
+  row_name <- match.arg(row_name)
+
+  columns_string <- columns_string_d2(
+    attrs,
+    attr_labels,
+    keys,
+    classes
+  )
+  columns_label <- columns_string
+  paste0(
+    label,
+    ": {",
+    "\n",
+    "  shape: sql_table",
+    "\n",
+    columns_label,
+    "\n}"
   )
 }
 
@@ -415,7 +599,28 @@ relation_schema_string <- function(
   )
 }
 
-columns_string <- function(col_names, col_labels, keys, col_classes) {
+relation_schema_string_d2 <- function(
+  attrs,
+  attr_labels,
+  keys,
+  name,
+  label
+) {
+  columns_string <- columns_schema_string_d2(attrs, attr_labels, keys)
+  columns_label <- columns_string
+  paste0(
+    "\"",
+    name,
+    "\": {",
+    "\n",
+    "  shape: sql_table",
+    "\n",
+    columns_label,
+    "\n}"
+  )
+}
+
+columns_string_gv <- function(col_names, col_labels, keys, col_classes) {
   key_membership_strings <- vapply(
     col_names,
     \(nm) paste(
@@ -440,6 +645,17 @@ columns_string <- function(col_names, col_labels, keys, col_classes) {
     key_membership_strings,
     "<TD PORT=\"FROM_", col_labels, "\">", col_classes, "</TD>",
     "</TR>",
+    recycle0 = TRUE
+  )
+  paste(column_typing_info, collapse = "\n")
+}
+
+columns_string_d2 <- function(col_names, col_labels, keys, col_classes) {
+  column_typing_info <- paste0(
+    "  \"",
+    col_names,
+    "\": ",
+    col_classes,
     recycle0 = TRUE
   )
   paste(column_typing_info, collapse = "\n")
@@ -476,6 +692,16 @@ columns_schema_string <- function(col_names, col_labels, keys) {
     "</TD>",
     key_membership_strings,
     "</TR>",
+    recycle0 = TRUE
+  )
+  paste(column_typing_info, collapse = "\n")
+}
+
+columns_schema_string_d2 <- function(col_names, col_labels, keys) {
+  column_typing_info <- paste0(
+    "  \"",
+    col_names,
+    "\"",
     recycle0 = TRUE
   )
   paste(column_typing_info, collapse = "\n")